Change related table-records
*Records
Sometime no data avialable in database so we must give warning or indication to user that no data available in database.
We done that changes in this commit.

*Font-awesome icon in table
We added trash and edit icon replacing the name delete and edit because it look good and requires less space compared to names